Extracted DelayScreen out of PINSpill - Renamed from DelayScreen to Delay - Added interface - Made helper functions for better readability
